Changes for the Style.css
I have added a lot of elements for CSS. The footer Has been 
centered.There's an input that controls "Write me a message". I was only 
able to have it control name and e-mail. I had to make a class called 
"message box" for the Textarea. 
I made another class for the Header contacts. I also did remove the 
header called "Contacts". I did not think it was necessary; moreover, I 
did move the contacts Field to a lower location underneath "Write me a 
message". That made it look more coherent.
I change the background of the 'main' class. The background is with a 
lighter blue. The text-fields are with a more dark light blue.
I changed the paragraphs to have a difference font-family. It follows a 
monospace font type.